requirements.txt
# Core Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
mamba-ssm>=1.2.0  # Requires CUDA

# Power System Simulation
pandapower>=2.13.0
networkx>=3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Optimization
cvxpy>=1.4.0  # For differentiable optimization layer
cvxpylayers>=0.1.6

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Jupyter (optional)
jupyter>=1.0.0
ipykernel>=6.23.0
